+++ scripts/jest-setup.js
@@ -1,0 +1,5 @@
+// @flow
+
+import {setTimezone} from '@frogpond/constants'
+
+setTimezone('America/Chicago')